requirements.txt
streamlit
yfinance
pandas
plotly
google-generativeai
python-dotenv
matplotlib
tabulate